+++ vacefron/__init__.py
@@ -1,7 +1,9 @@
 from .client import Client
 from .errors import *
 from .models import *
-from .enums import Badges, CrewmateColour
+from .enums import Badge, CrewmateColour
+
+Badges = Badge
 
 __license__ = "MLP-2.0"
 __author__ = "Soheab"